Move to logging.py module and add warning about multithread/process
